+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/guide-dog-service-dog/components/anonymous/steps-dt-details.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, ViewChild, ViewEncapsulation } from '@angular/core';
 import { ApplicationTypeCode } from '@app/api/models';
 import { BaseWizardStepComponent } from '@app/core/components/base-wizard-step.component';
 import { UtilService } from '@app/core/services/util.service';
-import { StepGdsdTermsOfUseComponent } from '../common-step-components/step-gdsd-terms-of-use.component';
+import { StepGdsdTermsOfUseComponent } from '../shared/common-step-components/step-gdsd-terms-of-use.component';
 
 @Component({
 	selector: 'app-steps-dt-details',
